+++ docs/CHANGELOG.md
@@ -16,7 +16,8 @@
 
 ### Components
 
-- [x] Icon component to render either google or custom icon
+- [x] Icon component to render custom icon
+- [x] Icon Material component to display google icons
 - [x] Message component to display information with severity (info, warn, error ..)
 - [x] Notification center (display notification in the root component)